// scrivafowcl/src/app/shared/models/formio/formio-imports-config.ts
/*
 * ========================LICENSE_START=================================
 * 
 * Copyright (C) 2025 Regione Piemonte
 * 
 * SPDX-FileCopyrightText: (C) Copyright 2025 Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
export interface FormioImportsConfig {
    GETISTANZA?: boolean;
    GETJSONDATA?: string[];
    GETSOGGETTI?: boolean;
    GETISTANZACOMPETENZAACP?: boolean;
    DOCSEGR?: boolean;
    DOCAMM?: boolean;
    GESTIONEPRIVACY?: boolean;
    GDPRDOCPATH?: string;
}
  